hw-06 refactor: rewrite dao layer to jpa (hibernate as JPA-provider)

+++ hw-06/src/main/java/ru/otus/booklibrary/services/BookService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface BookService {
 
-    Book getById(int id);
+    Book getById(long id);
 
     Book getByName(String name);
 
@@ -18,5 +18,5 @@ public interface BookService {
 
     Book save(Book book);
 
-    boolean deleteById(int id);
+    boolean deleteById(long id);
 }
